✨ useGetMovies
